SAK-51450 SiteStats replace custom SakaiPagingNavigator with standard Wicket 9 components
This commit refactors the pagination in SiteStats to use standard Wicket 9 components instead of custom ones:

1. Removed custom SakaiPagingNavigator.java and its HTML template
2. Updated SakaiNavigationToolBar to use the standard AjaxPagingNavigator component
3. Moved row size selector functionality to a separate component in SakaiNavigationToolBar
4. Updated SakaiNavigationToolBar.html template to include the row size selector element
5. Added null check to prevent duplicate component IDs when the toolbar is re-rendered

The refactoring preserves all the original functionality while removing custom code that is no longer needed with Wicket 9, making future maintenance easier.
