+++ sitestats/sitestats-tool/src/webapp/html/components/SakaiNavigationToolBar.html
@@ -26,9 +26,16 @@
             <div style="text-align: center;" class="listNav">
                 <div wicket:id="navigatorLabel" class="instruction">[navigator-label]</div>
                 <span wicket:id="navigator">[navigator]</span>
+                <select wicket:id="rowSizeSelector">
+                    <option>5</option>
+                    <option>10</option>
+                    <option>20</option>
+                    <option>50</option>
+                    <option>100</option>
+                    <option>200</option>
+                </select>
             </div>
         </td>
     </tr>    
 </wicket:panel>
 </html>
-
