--- sitestats/sitestats-tool/src/webapp/html/components/SakaiPagingNavigator.html
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!--
-
-    $URL:$
-    $Id:$
-
-    Copyright (c) 2006-2009 The Sakai Foundation
-
-    Licensed under the Educational Community License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-                http://www.osedu.org/licenses/ECL-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<html xmlns:wicket>
-<body>
-  <wicket:panel>
-    <!-- <a wicket:id="first">&lt;&lt;</a>&nbsp;<a wicket:id="prev">&lt;</a> -->
-    <input type="submit" wicket:id="first" wicket:message="value:pager_textFirst"/>
-    <input type="submit" wicket:id="prev" wicket:message="value:pager_textPrev"/>
-    <span wicket:id="navigation">
-          <a wicket:id="pageLink" href="#"><span wicket:id="pageNumber">5</span></a>
-    </span>
-    <!-- <form wicket:id="form"> -->
-	    <select wicket:id="rowNumberSelector">
-	        <option>100%</option>
-	        <option>10%</option>
-	    </select>
-    <!-- </form> -->
-    <input type="submit" wicket:id="next" wicket:message="value:pager_textNext"/>
-    <input type="submit" wicket:id="last" wicket:message="value:pager_textLast"/>
-    <!-- <a wicket:id="next">&gt;</a>&nbsp;<a wicket:id="last">&gt;&gt;</a> -->
-  </wicket:panel>
-</body>
-</html>
